requirements.txt
boto3==1.26.123
botocore==1.29.123
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
dnspython==2.3.0
h3==3.7.6
idna==3.4
iso3166==2.1.1
jmespath==1.0.1
langdetect==1.0.9
msgpack==1.0.5
numpy==1.24.3
origamibot==2.3.4
polyline==2.0.0
pyaes==1.6.1
pyasn1==0.5.0
pycparser==2.21
pymongo==4.3.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.29.0
rsa==4.9
s3transfer==0.6.0
six==1.16.0
telegraph==2.2.0
Telethon==1.28.5
timezonefinder==6.2.0
uritemplate==4.1.1
urllib3==1.26.15
